+++ src/main/java/frc/robot/RobotContainer.java
@@ -8,6 +8,8 @@ import com.ctre.phoenix6.Utils;
 import com.ctre.phoenix6.swerve.SwerveRequest;
 import com.pathplanner.lib.auto.AutoBuilder;
 
+import edu.wpi.first.math.geometry.Pose2d;
+import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
@@ -16,6 +18,7 @@ import edu.wpi.first.wpilibj2.command.button.RobotModeTriggers;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
 import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine.Direction;
 import frc.robot.commands.DriveCommand;
+import frc.robot.commands.AutoDrive;
 import frc.robot.subsystems.SuperStructure;
 import frc.robot.subsystems.climber.ClimberIO;
 import frc.robot.subsystems.climber.ClimberIOSim;
@@ -137,7 +140,13 @@ public class RobotContainer {
             )
         );
 
-        // driverController.a().onTrue(new AutoDrive(drivetrain).withTarget(new Pose2d(0, 0, Rotation2d.fromDegrees(0))));
+        // TODO drivetest
+        driverController.a().onTrue(new AutoDrive(drivetrain).withTarget(new Pose2d(5, 5, Rotation2d.fromDegrees(0))));
+        driverController.x().onTrue(
+            drivetrain.runOnce(
+                () -> drivetrain.resetPose(new Pose2d(2.5, 4.5, Rotation2d.kZero))
+            )
+        );
         //#endregion
     }