+++ src/main/java/frc/robot/subsystems/drive/ChassisConstants.java
@@ -7,18 +7,30 @@ import edu.wpi.first.math.geometry.Rotation2d;
 
 public class ChassisConstants {
     //#region modifiable
-    public class SIMULATION {
+    private class SIMULATION {
         private static final double TRANSLATION_PID_P = 5.0;
         private static final double TRANSLATION_PID_D = 0.0;
         private static final double ROTATION_PID_P = 15.0;
         private static final double ROTATION_PID_D = 0.0;
     }
 
-    public class REAL {
-        private static final double TRANSLATION_PID_P = 5.0;
+    private class REAL {
+        private static final double TRANSLATION_PID_P = 0.0;
         private static final double TRANSLATION_PID_D = 0.0;
-        private static final double ROTATION_PID_P = 2.0;
+        private static final double ROTATION_PID_P = 0.0;
         private static final double ROTATION_PID_D = 0.0;
+    }
+
+    private class AUTOAIM_ROTATION{
+        private class SIMULATION {
+            private static final double ROTATION_PID_P = 15.0;
+            private static final double ROTATION_PID_D = 0.0;
+        }
+
+        private class REAL{
+            private static final double ROTATION_PID_P = 5.0;
+            private static final double ROTATION_PID_D = 0.0;
+        }
     }
 
     public static final Pose2d TOLORANCE_AUTO_DRIVE = new Pose2d(
@@ -28,7 +40,7 @@ public class ChassisConstants {
     );
     //#endregion
 
-    //#region computations
+    //#region Pathplanner & AutoDrive
     public static final double TRANSLATION_PID_P = 
         Utils.isSimulation() 
             ? SIMULATION.TRANSLATION_PID_P 
@@ -48,5 +60,17 @@ public class ChassisConstants {
         Utils.isSimulation()
             ? SIMULATION.ROTATION_PID_D
             : REAL.ROTATION_PID_D;
+    //#endregion
+
+    //#region AutoAim
+    public static final double AUTOAIM_ROTATION_PID_P = 
+        Utils.isSimulation() 
+            ? AUTOAIM_ROTATION.SIMULATION.ROTATION_PID_P 
+            : AUTOAIM_ROTATION.REAL.ROTATION_PID_P;
+
+    public static final double AUTOAIM_ROTATION_PID_D =
+        Utils.isSimulation()
+            ? AUTOAIM_ROTATION.SIMULATION.ROTATION_PID_D
+            : AUTOAIM_ROTATION.REAL.ROTATION_PID_D;
     //#endregion
 }