+++ changelog.md
@@ -9,6 +9,14 @@ date - post `commit ID (short version)`
 
 # Todo
 
+## post `97d14c3`
+- a KidsMode Axis Config for swerve
+- catch BufferUnderFlow for Logging
+- Fix Rotation Feedforward for AutoAim
+- Some changes that will be revert after drivetest
+- Different Constants for AutoAim Rotation PID
+- rough tune for AutoAim Rotation PID Controller   
+
 ## post `d1a0688`
 - Moved Climber Voltage to the Constant File
 - format comments